[STORM-311] Disallow override of runner parameters in action.
- At registration actions are disallowed from overriding the parameters
  defined by corresponding runners.
- Moved ActionType from the st2actionrunnercontroller project to
  st2actioncontroller project. This way ActionTypes are queriable from
  the same pecan controller as /actions.
- Moved ActionType registration out of controller code to have the
  registration become more explicit.
- Rearranged how ActionTypes and Actions are registered.
