+++ st2actioncontroller/st2actioncontroller/cmd/actioncontroller.py
@@ -1,8 +1,6 @@
 import eventlet
 import os
 import sys
-import glob
-import json
 
 from oslo.config import cfg
 from wsgiref import simple_server
@@ -12,8 +10,10 @@ from st2common.models.db import db_setup
 from st2common.models.db import db_teardown
 from st2actioncontroller import config
 from st2actioncontroller import app
-from st2common.persistence.action import Action
-from st2common.models.db.action import ActionDB
+from st2actioncontroller import model
+
+
+from wsgiref import simple_server
 
 
 eventlet.monkey_patch(
@@ -42,23 +42,11 @@ def __setup():
     if not os.path.exists(cfg.CONF.actions.modules_path):
         os.makedirs(cfg.CONF.actions.modules_path)
 
-    # 5. register actions at the modules path
-    actions = glob.glob(cfg.CONF.actions.modules_path + '/*.json')    
-    for action in actions:
-        with open(action, 'r') as fd:
-            content = json.load(fd)
-            try:
-                model = Action.get_by_name(str(content['name']))
-            except:
-                model = ActionDB()
-            model.name = str(content['name'])
-            model.description = str(content['description'])
-            model.enabled = bool(content['enabled'])
-            model.artifact_paths = [str(v) for v in content['artifact_paths']]
-            model.entry_point = str(content['entry_point'])
-            model.runner_type = str(content['runner_type'])
-            model.parameters = dict(content['parameters'])
-            model = Action.add_or_update(model)
+    # 5. register actiontypes and actions. The order is important because actions require action
+    #    types to be present in the system.
+    model.register_action_types()
+    model.register_actions()
+
 
 def __run_server():
