+++ st2actioncontroller/st2actioncontroller/controllers/root.py
@@ -5,7 +5,7 @@ from webob.exc import status_map
 from st2common import log as logging
 from st2actioncontroller.controllers.actions import ActionsController
 from st2actioncontroller.controllers.actionexecutions import ActionExecutionsController
-
+from st2actioncontroller.controllers.actiontypes import ActionTypesController
 
 LOG = logging.getLogger('st2actioncontroller')
 
@@ -15,6 +15,8 @@ class RootController(object):
     actions = ActionsController()
     # Handler for /actionexecutions/
     actionexecutions = ActionExecutionsController()
+    # Handler for /actiontypes/
+    actiontypes = ActionTypesController()
 
 # TODO: Remove index handler
     @expose(generic=True, template='index.html')
