+++ st2actioncontroller/tests/controllers/test_actiontypes.py
